initial commit: CAT driven LZ2RR web switch
This is a CAT driven automatic band switcher that interfaces
with the LZ2RR web based antenna switch(es).

It connects to a serial port and is tested with a kenwood ts590sg,
but in theory should work with most transceivers where the protocol is similar to Kenwood.
At least in this case all it cares about are the IF and FA commands.